Employee-portal/requirements.txt
# Django & DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.4
python-decouple==3.8
django-celery-beat==2.6.0

# JWT Authentication
PyJWT>=2.8,<3

# AI & ML (pinned for compatibility)
openai>=1.30,<2
langchain>=0.2,<0.3
langchain-core>=0.2,<0.3
langchain-openai>=0.1,<0.2

# Vector DB / embeddings
chromadb>=0.5.0,<0.6

# Env
python-dotenv==1.0.0

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11
Pillow==10.1.0

# Celery & Task Queue
celery==5.3.4
redis==5.0.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Development Tools
Django-debug-toolbar==4.2.0
black==23.12.0
flake8==6.1.0
isort==5.13.2

# Logging and Monitoring
python-logging-loki==0.3.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
requests==2.31.0
pytz==2023.3
